refator numpy api
